+++ test-build.sh
@@ -24,13 +24,8 @@ cargo build --no-default-features --features managed,unmanaged
 	cargo build
 	cargo build --no-default-features --features rt_tokio_1
 	cargo build --no-default-features --features rt_async-std_1
-)
-
-(
-	cd redis-cluster
-	cargo build
-	cargo build --no-default-features --features rt_tokio_1
-	cargo build --no-default-features --features rt_async-std_1
+	cargo build --no-default-features --features rt_tokio_1 cluster
+	cargo build --no-default-features --features rt_async-std_1 cluster
 )
 
 (