+++ redis/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v0.14.0
+
+* Merge `deadpool-redis-cluster` into `deadpool-redis`.
+* Remove `redis_cluster_async` dependency in favor of `redis::cluster` / `redis::cluster_async`.
+
 ## v0.13.0
 
 * Update `deadpool` dependency to version `0.10`